Fix MessageDelivery entity by changing ID types to int
Updated MessageId and ChatId types in MessageDelivery from long to int for consistency across the database. Adjusted related migrations and foreign key relationships to reflect the changes.

+++ src/TelegramBot.ApiService/Domain/Entities/MessageDelivery.cs
@@ -2,8 +2,8 @@
 
 public class MessageDelivery : BaseEntity
 {
-    public long MessageId { get; set; }
-    public long ChatId { get; set; }
+    public int MessageId { get; set; }
+    public int ChatId { get; set; }
     public DeliveryStatus Status { get; set; }
     public string? ErrorMessage { get; set; }
     public DateTime? SentAt { get; set; }
